Assignment 3
Q1. Which keyword is used to create a function? Create a function to return a list of odd numbers in the range of 1 to 25.
Q2. Why *args and **kwargs is used in some functions? Create a function each for *args and **kwargs to demonstrate their use.
Q3. What is an iterator in python? Name the method used to initialise the iterator object and the method used for iteration. Use these methods to print the first five elements of the given list [2, 4, 6, 8, 10, 12, 14, 16, 18, 20].
Q4. What is a generator function in python? Why yield keyword is used? Give an example of a generator function.
Q5. Create a generator function for prime numbers less than 1000. Use the next() method to print the first 20 prime numbers.