Assignment 6
Q1, Create a vehicle class with an init method having instance variables as name_of_vehicle, max_speed and average_of_vehicle.
Q2. Create a child class car from the vehicle class created in Que 1, which will inherit the vehicle class. Create a method named seating_capacity which takes capacity as an argument and returns the name of the vehicle and its seating capacity.
Q3. What is multiple inheritance? Write a python code to demonstrate multiple inheritance.
Q4. What are getter and setter in python? Create a class and create a getter and a setter method in this class.
Q5.What is method overriding in python? Write a python code to demonstrate method overriding.