Exception handling-1
Q1. What is an Exception in python? Write the difference between Exceptions and Syntax errors.
Q2. What happens when an exception is not handled? Explain with an example.
Q3. Which Python statements are used to catch and handle exceptions? Explain with an example.
Q4. Explain with an example:#
 a. try and else
b. finally
c. raise
Q5. What are Custom Exceptions in python? Why do we need Custom Exceptions? Explain with an example.
Q6. Create a custom exception class. Use this class to handle an exception.