MongoDB
Q1. What is MongoDB? Explain non-relational databases in short. In which scenarios it is preferred to use
MongoDB over SQL databases?
Q2. State and Explain the features of MongoDB.
Q3. Write a code to connect MongoDB to Python. Also, create a database and a collection in MongoDB.
Q4. Using the database and the collection created in question number 3, write a code to insert one record,
and insert many records. Use the find() and find_one() methods to print the inserted record.
Q5. Explain how you can use the find() method to query the MongoDB database. Write a simple code to
demonstrate this.
Q6. Explain the sort() method. Give an example to demonstrate sorting in MongoDB.
Q7. Explain why delete_one(), delete_many(), and drop() is used.